requirements.txt
base36==0.1.1
python-dotenv==1.0.0